more work on multiprocessing
a lot of work and reading on the internet, not a lot of progress though. nevertheless, i need to get this to work because I bet as a team a lot of our work involves multiprocessing, so this is good for my coding development so i can use multiprocessing comfortably in the future.

hack does not work (initially at least)...https://gist.github.com/EdwinChan/3c13d3a746bb3ec5082f

maybe best bet for now is to try to get it to work though.

somethign to look into is where is multiprocessing most efficient and where in code does it deliver biggest speed ups (running one function multiple times, or parallelizing aspects of that function and running the function itself sequentially for different files)