post prediction fix doesn't work; back to debugging
approach for today: invert at end by saving min and max, then turning prediction matrix A' = 1/A, then renormalizing A' between those boundaries.
I'm going to make the change after the detect trees step of predict() because it's the least involved aspect of the code to change.

X_std = (X - X.min(axis=0)) / (X.max(axis=0) - X.min(axis=0))
X_scaled = X_std * (max - min) + min

Okay based on initial investigation this does not seem like a good approach...it reverses the distribution which is a bad idea.

maybe my norm func is wrong?

maybe 2d norm is bad? haha lol it was using l1 normalization...switched to max
axis was problem because it was normalizing by row in the matrix (i.e. by sample)

okay so I've proved that there's the inversion because if you divide 1 by the output and renormalize (and adjust th) you get polygons out...this is not long term fix however (as we discussed)...because the activation confidence could get messed up by inversion, such that if it wasn't inverted max confidence would be 0.9 for tree,  but because of weirdness introduced from inversion, the max confidence for background detection would be lower (or greater) like 0.7.
This isn't even to mention how janky it would be to mess with the threshold (th) for polygon detection on the renormalized prediction mask

priority is to find and fix the root culprit bug before the weekend ends!!

immediate to-do: migrate plots-for-debugging out of this repo and fix main.py and prediction.py