breaking ground on pytorch
hold on tight folks, this is going to be a crazy ride.... :horse_racing: 🤠🤠

first attempt, just take out the model.train and output steps.

might need to change data prep steps later.

okay...after cloaning in two separate UNET implementations in pytorch, this is actually going to be a lot more intense than I anticpated haha

plan is to try to get some version working asap while spending time learning 1. on the side from scratch from that book i got online; and 2. on the fly trying to get this first version to work.

immediate things to work on are loading our data into pytorch data format for training and predictions

will definitely need to inspect this model to make sure it is identical in architecture and stuff like loss and optimization function stuff as our tensorflow one.