implemented patch sampling
Questions:
1. normalizing patches (40% 60%...from paper)
2. image augmentation? not needed yet? i can't see how most would help?

notes

#
# https://machinelearningmastery.com/difference-between-backpropagation-and-stochastic-gradient-descent/
#
# https://stackoverflow.com/questions/53975717/pytorch-connection-between-loss-backward-and-optimizer-step
# https://www.tomasbeuzen.com/deep-learning-with-pytorch/README.html

So, when we call loss.backward(), the whole graph is differentiated w.r.t. the loss, and all Variables in the graph will have their .grad Variable accumulated with the gradient.

want to be really careful when defining the loss, because I don't want to calculate it wrong.

what is the computational graph leading to the associated n-dimensional matrix in torch tensors?

need to review a lot of multi/linear alg/diffy Qs this summer.

I didn't need to have backward() method or make my tversky_loss function a class with inheritance (tensor has the method backward() in itself)