Add enhanced vector preprocessing methology
preprocessing.py suffers from over-complexity and O(N^2) algorithms.  The problem can be recast in terms of raster operations (as opposed to geospatial database operations) for big wins in simplicity and performance.

The output of this script does not yet observe the data convention assumed by the training process.  The boundary and annotation masks are interwoven in the same band (0 == backgrounnd, 1 == positive training data, 2 == overlap boundary), instead of breaking annnotations and boundaries into separate files.

We do not want a proliferation of files grouped by ascii numbers in their filenames when they are conceptually part of the same parcel of data.  So this this one possible choice of improvement on the data preparation side.