Change stage 0 cutout count to 10, and other minor simplifications
The script will now throw a fit if imported elsewhere.  We can always graft this in as a library for re-use later, but we don't have that problem today.

The cutout profile handling is now more explicit and in your face about its configuration.  Otherwise I found it too easy to gloss over the myriad of rasterio.open parameters.

TODO: The overlap bug still exists!