Add stage 4: Predict
This script applies a trained CNN to an entire mosaic tile.

It also transfers the no-data regions from the tile to the output in such a way that positive predictions over no-data are still preserved.

Currently, this assumes about ~6G of memory are available on the machine.  You are free to make local changes to this script to do window reads, or process a smaller subset of a tile, in the event your machine is under-powered, but such contributionos are be forbidden in the repository proper.