Add stage 4: tree heights determination UNTESTED
This commit adds a 4th stage that transforms the CNN's shadow raster detection into polygons that, when paired with the associated cutline geometry, are used to determine the shadow length and tree height using the solar / view geometry.

The shadow length / height estimation has not been implemented and is marked as TODO.  All the prior labor so far is in service to that tiny little bit of yet to be written code.