Finish initial stage 4 testing, and introduce stage 5.
Stage 4 is missing the critical component of actually determining the shadow length and associated tree height, but the context has now been established for further development.

Stage 5 now exists, ta da, but in TODO form.  Determining tree heights alone is publishable, but I suspect a top tier journal publication if we can combine the heights with area to better quantify total carbon in the study region.