UNTESTED. Stage 4: Coregister shadow polygons with matching tree canopy, determine shadow length and calculate tree height + carbon estimation
This is very TODO, incomplete, but I wanted to commit what changes I had before the weekend hit because Thaddaeus is back.

This script is the 'hard' part of the problem.  And I've only touched the surface.

I've implemented, probably w/ bugs, Arthur's shadow length proposal with some enhancements and simplifications.  What makes it tricky is that the geospatial coordinate system is at odds with conventional mathematics.  So very special care must be used when inputing trig functions with angles from geospatial sources.  Then there's the linear algebra, which seems to very neatly fold down to a simple dot product.  We basically construct a vector from a matched tree crown centroid to a vertex in the shadow geometry and measure how much of the x,y components are aligned with the solar flux direction.  The largest such contribution is considered the shadow length.

Dr Tucker sent me Pierre's allometry equations for crown height * crown area carbon stock determination, which are implemented.

The script doesn't *do* anything with the heights or lengths or carbon yet.