Stage 0: Generate all patches & sample chirps,  Stage 3: Special case NAN in standardization
Stage 0 now generates cutouts given the entire corpus of mosaic tiles across all UTM zones.  It also adds a CHIRPS metadata item, which measures the annual rainfall in the area.  The "sample the Chirps raster from the cutout bounds" code is unfortunately a bit complex because of how the various packages interact.  I would expect rasterio to have a high level function for this, but alas, I have not found one.

Stage 3:  Don't make the nan case the common case.  We now check for no-data regions, and special case the nan handling.  Turns out by doing this we bypass the need for is_nan() because we already know where the no-data values are.