Stage 2: Test Updated UNet
Set UpSampling2D kernel to bilinear to allow GPU kernel execution
With float16, we can up the batch count to 96 for much improved stability during training.  On NCCS we can bump this even higher on their A100s.
The boundary weight needed to be either 1 or 10, not 0 or 10.  Fixed.
Frame image names are now serialized to disk (instead of their useless indices).
adaDelta seems to edge out over adam as an optimizer, but later we can generate a mix of UNets and see which is best.