Update Stage 3 to the new UNet
Follow the recommendation from Keras to use predict_on_batch as opposed to predict for our use case.  It saves about 20 seconds from the predict segment.  The cost is that the indexing is a little more complicated to track how predictions map to the output array as they are now decoupled.

Also refine the sozip logic a little by removing paths from the archive name, and later removing temporary files.